Move multiplexer name to toys.h
It can be used to write build scripts to either apply patches to toys.h or change the name directly there. But since its out-of-order ie., the Makefile and other build scripts have hard-coded this name, it's necessary to rename the output binary as well ie., the output still produces toybox-prefixed binary and it is necessary to rename the prefix to whatever the TOYBOX_NAME is set.